requirements.txt
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
transformers==4.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
librosa==0.10.1
opencv-python==4.8.1.78
facenet-pytorch==2.5.3
pyyaml==6.0.1
